add `/building` endpoint
This endpoint return buildings object.
It add two other fields:
    - `surface` the total surface of the building
    - `most_usage` the most usage of the buiding

`surface` is computed by suming all building's zones surface

The `most_usage` is the usage that is most present in zones.
To have it, we start by summing all usages surface, by suming zones by
usages. Next we get the greatest of all usage surface, and return the
usage belonging to it