feat: add building carbon impact computation
The carbon impact of a building is the sum of the carbon impacts of all
its zone.

Each of these zone carbon impacts are computed by suming carbon
impacts of zone's construction products.

In each construction products, we have 4 impacts types:
    - The production impact
    - The construction impact
    - The exploitation impact
    - The endoflife impact

In this commit we only manage production, construction and endofife
impact. They are computed by multiplying the quantity of this products
in the zone, and the specific impact index in the products.
For example, in a zone A, with X quantity of the Product P, with the
impact index I, we have the carbon impact of the zone with X * I.
This rules in the same with all three impact types, just with their
specific index